More type issues

+++ editor/src/editor/panels/editor_panel_fonts.cpp
@@ -62,7 +62,7 @@ void EditorPanelFonts::DrawContents() {
     if (ImGui::BeginListBox("##font_list", ImVec2(-FLT_MIN, 0))) {
         for (size_t i = 0; i < fontNames.size(); ++i) {
             auto const& entryInfo = fontNames[i];
-            bool const selected = m_selectedIndex == i;
+            bool const selected = static_cast<size_t>(m_selectedIndex) == i;
             if (ImGui::Selectable(entryInfo.c_str(), selected)) {
                 m_selectedIndex = i;
             }
